# perf(test): use fast ops for deno test register